feat(sam): rename topological sorting related interfaces
A topological ordering of states is not necessarily
an ordering from suffix parents to their children.
So an arbitrary topological ordering cannot be used
for updating accepting states.

However, the implementation is still correct.
Because the topological sorting is done with a queue,
where **NO** states with longer maximum suffix lengths
are visited earlier than states with shorter ones.

So the interface names related to topological sorting
are changed into more precise ones.